Enable reading /etc/rhsm-conduit.conf
To test, create a file: /etc/rhsm-conduit.conf with
hostInventoryServiceUrl set to foobar and then launch the jar;
observe the message:

```
host inventory service URL: foobar
```

Also, this adds a factory for the host inventory client that is
constructed using this config key. This pattern can be used in other
other swagger clients.

With the factory in place, it is sufficient to @Autowire in a ApiClient,
as seen in ApiClientTest.