Enable checkstyle for inventory and pinhead subprojects.
The inventory and pinhead subprojects are hybrids of generated and
authored code.  Both types of code are in the same Gradle source set.
We need to run style checks on the authored code and I have not seen any
issue pop up with the generated code, in the future we may need to
separate them into different source sets to keep style checking working.