Created a task framework for performing asynchronous work
=== Usage ===
All that should be required to use this framework is to autowire a TaskManager where applicable
and to call one of its methods to trigger an async task.

@Autowired
private TaskManager tasks;

// In your method, use the manager to queue up a task.
tasks.updateOrgInventory(orgId);

=== Framework Details ===

This framework is made up of a few key components;

TaskManager - All task queuing is done though methods on this class

TaskDescriptor - Describes a Task that is to be executed. In the case of Kafka, this will
                 describe the message that is to be stored.

TaskQueue - The TaskQueue interface provides a means to implement various queuing implementations.
            It stores TaskDescriptors that are to be converted to Task objects by the TaskWorker
            when it executed.

TaskWorker - Builds a Task from a TaskDescriptor and executes the task.

Task - Represents the actual work that needs to be done.

=== Configuration ===
All available TaskQueue configuration is done based on a configuration
property.

rhsm-conduit.tasks.queue=<QUEUE_KEY>

By default, a **pass-through** queue is used. This queue simply passes a queued TaskDescriptor
on to the TaskWorker to be executed immediately.

When other queue implementations are created, an application configuration should be created for
this queue and all beans should be defined with the appropriate @ConditionalOnProperty annotation
that filters on a custom QUEUE_KEY for this type of queue.

@Bean
@ConditionalOnProperty(prefix = "rhsm-conduit.tasks", name = "queue", havingValue = "my-cool-queue")
TaskQueue passThroughQueue() {
    log.info("Configuring my cool task queue.");
    return new CoolQueue();
 }

Alternatively, if we'd like to keep all queue configurations in the same place, we could define
the queue related beans in the TaskQueueConfiguration class with the appropriate conditional
annotations.