Used Spring Kafka to implement a Kafka TaskQueue implementation.
When the TaskManager queues a task, a message is sent to Kafka
via the rhsm-conduit-tasks topic.

A KafkaListener is configured to listen to the rhsm-conduit-tasks topic
and is notified when a message is available. The message is then translated
to a Task and is run via the TaskWorker.

If Task execution fails, the error is logged and the message is considered
processed.

To enable this queue, the following property must be set as an env var
or in the rhsm-conduit.properties config file:
    rhsm-conduit.tasks.queue=kafka

Spring Kafka's auto-configure is used to configure the producer/consumer. Any of the
confurations defined by Spring Kafka are defines with the "spring.kafka" prefix as
per the documentation. Customizing the consumer or producer via properties not included
with auto-configure can be specified via:
   spring.kafka.consunmer.properties.MY_CONSUMER_PROPERTY
   spring.kafka.producer.properties.MY_PRODUCER_PROPERTY

To increase the number of messages that can be processed concurrently, the
topic must be configured with a number of partitions equal to the value of the
spring.kafka.listener.concurrency property.