Wire Quartz up with a JDBC backed data store
This patch adds support for either a PostgreSQL or HSQLDB data store for
Quartz.  Depending on the setting of rhsm-conduit.datasource.platform,
Spring will create the schema in the selected database implementation.
By default, an HSQLDB database will be used, but if the properties
packaged within the application are overrided externally, PostgreSQL can
be used.