Implement ExecutorTaskQueue/in-memory task queue
See javadoc of `ExecutorTaskQueue` for details. The idea is that by
implementing a different `BlockingQueue`, we'll be able to integrate with
Kafka (or any other messaging).

`TaskQueueConfiguration` is now responsible for configuring the behavior
of the `ExecutorTaskQueue`. Future work could provide options to
configure the number of threads or bounds on the queue.