Implement the debug endpoint
With this change the debug endpoint accurately reflects what would be
sent to the inventory service.

Since we control the interface for it, I decided that ConsumerInventory
should reflect our service's view of a consumer generally, and then
realized that there was a great deal of overlap with ConduitFacts.

Since the validation provided by ConduitFacts has been valuable, I
reworked it to be a subclass of ConsumerInventory.

With this in place, `curl localhost:8080/rhsm-conduit/v1/inventories/1234`
(or similar) is useful.