Create normalized facts from inventory host FactSet
Any fact that should be included in the normalized facts should
have an associated FactSetEnricher that will pull the value from
the appropriate inventory FactSet namespace and set it on the passed
EnrichedFacts object.

FactSetEnricher instances are set on the Enricher and will process
all FactSet objects returned from the inventory for a host.