Convert entire project to use dependency management.
Using the dependency management plugin allows us to impose regimented
restrictions on our dependency versions.  A dependency is defined in the
dependency management section and then specified without a version when
it is listed as a compile/test/runtime dependency.  This system is more
succinct than using version variables and it allows us to consume Maven
Bill of Materials POMs where we can draw from a list of versions blessed
in a Spring release.

See
https://docs.spring.io/dependency-management-plugin/docs/current-SNAPSHOT/reference/html/