Add pagination information; remove account number from API.
Insights has a standard for responses that we are supposed to follow.
That standard includes allowing for pagination in requests and returning
a block of navigation links in responses.

This commit also removes the account number as a path parameter from the
API.  The account number that is delivered via the x-rh-identity header
is the only account the API will service.