Fixed missing transaction issue when purging snapshots
Spring does not honor @Transactional when the method it is on
is called from within the same class. In this particular case,
we would have to mark purgeSnapshots to make it work. Since we
want a transaction around the purge for each account, the
annotation was placed on the repository method itself.