Add subscription_capacity table
The intention for a given subscription, which can be identified by a
subscription ID, to collect each applicable product's contributing
socket count, and whether the subscription provides unlimited guests
(VDC-style SKUs).

This does not handle any local caching of product data, which we may
elect to do in the future in order to improve performance.

This table is intentionally denormalized, as we intend to start out by
using denormalized views of subscription data from Candlepin pools.