Add ingress API and implementations
The idea is that we will have at least two deployment types (think
subscriptions vs. subscriptions-jobs deployment configs). In the general
API case, the API will not be present. In the other case, we'll
configure the server to require client cert auth (and actually the
endpoint).

To test the stub endpoint, set `rhsm-subscriptions.enableIngressEndpoint=true`
in `config/application.properties`