Classify systems using a proxy repository.
In order to mark systems as hypervisors, we must look at all guests and
collate the hypervisor UUIDs they report cross-referenced with the
subscription manager IDs of reported systems.  We can then mark systems
as hypervisors and also flag guests who have hypervisor UUIDs that we
don't find in the entire system catalog.