Skip host in RHEL tally when is hypervisor and has no sockets
Since guests are skipped in the RHEL tally when the hypervisor
is known (data taken from hypervisor for calcs), if the hypervisor
has no sockets, then nothing will be tracked in the totals.

This could potentially happen in cases where virt-who is reporting
the host/guest mapping to candlepin and the host socket counts are
not known.