Use orgId over accountNumber.
In certain cases, a pool may be missing an account number.  For example,
if a customer acquires a developer subscription and then buys from Red
Hat, the developer subscription won't be updated with their account
number.  Consequently, relying on the account number as a component
of a primary key in a table is undesirable.

This patch changes the capacity table to use the orgId (which should
always be present) as a part of the composite primary key.
Additionally, it ties the principal to the orgId from the authentication
header and performs necessary queries with it.