Synchronize on the fact queue before sending
Since the inventory service is a singleton, facts from many tasks
can get put on the queue. We need to syncronize the use
of the task queue list queue while the messages are
being flushed. Creating a synchronized list is not enough
since we shouldn't allow the list to be modified until
the flush is complete.