Add the sla query parameter and include it in the tally response
Query used by API now does sums values on the DB server. This
was done to limit the amount of data that gets post-processed
which admittedly adds complexity but should reduce the memory
footprint when using the API.

An optional API parameter (sla) has been added to limit results
to snapshots with a matching sla only.

We have opted to add a magic value for the sla parameter ('unset')
that indicates that we are filtering on any snapshot that has no
SLA set.This facilitates the requirement of being able to query
for snapshots that have an empty SLA. Not setting a value for the
query parameter drives the "don't care" semantics of the underlying
query.