Add SLA to tally snapshots
I added an enum: `org.candlepin.subscriptions.db.model.ServiceLevel`
which represents all possible values, including special values
`UNSPECIFIED` and `ANY`. `ANY` is used to collect all SLAs into Snapshot
records.

I renamed `ProductUsageCalculation` to `UsageCalculation` as introducing
SLA as a discriminator made it so that UsageCalculations are keyed on
the (product, SLA) tuple.

I also modified the table to make the SLA column non-nullable and
defaulted existing data to `_ANY`.  FYI, the changeset on a test table
w/ ~3.6 million rows took ~233 seconds to complete.

I made testing changes to try to keep it easy to add more fields to the
tuple in the future.