Ignore hosts that have been culled in HBI
Any host that has been culled does not get included in the produced
snapshots. A host's culled timestamp is determined via the host's
stale_timestamp as per HBI's formula:

   culled_timestamp = stale_timestamp + <offset_in_days>

The application's configuration allows us to tune <offset_in_days>
in case the HBI's above changes later. The default offset is 14 days.

NOTE:
This PR filters by stale_timestamp at the query level as well as at
the processing level. This is done so that should we move away from
querying the HBI DB directly, it will will not get missed. It also
acts as a safe guard.