Refactor into spring profiles
The following profiles now exist:

- api: instance should service API requests
- capacity-ingress: instance should accept capacity records
- scheduler: instance should schedule tasks
- worker: instance should pick up tasks from kafka and run them, implies kafka-queue
- in-memory-queue: instance should queue and run tasks in-memory
- kafka-queue: instance should queue tasks to kafka, but doesn't necessarily run them

If you run `./gradlew bootRun`, it will default to:

- api
- capacity-ingress
- scheduler
- in-memory-queue

Below commands show some ways to set profiles, and the combinations I think we need.

```sh
./gradlew bootRun --args=--spring.profiles.active=api  # accept requests and consume tasks from kafka
./gradlew bootRun --args=--spring.profiles.active=capacity-ingress  # accept capacity updates
./gradlew bootRun --args=--spring.profiles.active=scheduler,worker  # schedule tasks and work them
./gradlew bootRun --args=--spring.profiles.active=worker  # pick up tasks from the kafka queue
```

Note: I don't see an immediate use for kafka-queue, but if we ever want to deploy a node which is only a producer, it would be the profile to use (along with scheduler).
Note: we will need to override profile for each deployment type in the openshift deployment configs.

Scope TaskManager/TallyJmxBean to worker and queue profiles, so that if the node is not handling work, it doesn't need a Kafka connection.

Set initial offset commit to "earliest", so that when we first start the consumers, they start with the earliest messages.