Add DB schema and repository for AdminConfig (opt-in)
A AccountConfig entry represents an account/org pair and
the functionality that is allowed for that pair. For example,
should data sync be enabled in tally/conduit? Or is reporting
enabled for that account in tally.

The intent here is to allow a single record to enable the
various configurable bits across both tally and conduit.