Add last_checkin_after argument to request to Pinhead.
Subtract the threshold on the last sync time from the time that we send
to Pinhead as the last_checkin_after parameter.

Remove last check date filtering now that it has moved up to Pinhead.

Do not use ISO_INSTANT datetime format as it can have nanoseconds.

Also, do not call the method annotated with validation annotations from
within the class.  Spring AOP proxies require that methods be called
externally from their own class in order for the AOP advice to trigger.