Prevent the security checks from happening multiple times per filter
When the CSRF filter was added, security checks were being performed
multiple times per request, resulting in processing the identity header
multiple times. This was happening because we were setting the decoded
header as the principal and then replacing it with the InsightsUserPrincipal
once the header JSON was was processed. Spring security was checking
to see if the principal had changed between filters and because it was
replaced, it checked it twice.

This PR does the header parsing and principal generation in the
IdentityHeaderAuthenticationFilter. This is in turn passed to the
IdentityHeaderAuthenticationManager which simply validates the data
in the principal and throws appropriate exceptions which trickles
back to the response with the appropriate codes.

By doing this, we do not swap the principal in the context, avoiding
the failed principal equality checks resulting in the extra security
checks.