Reconfigure Spring Security to use only one AuthenticationManager
Previously we had Spring Security configured to use one
AuthenticationManager in the IdentityHeaderAuthenticationFilter and then
another AuthenticationManager (ProviderManager) configured that
delegated to the PreAuthenticatedAuthenticationProvider.  That was not
correct.  There only needs to be one AuthenticationManager.

This patch creates a single ProviderManager which delegates to an
IdentityHeaderAuthenticationProvider (other providers can be added later
if desired).  This ProviderManager is given to the
IdentityHeaderAuthenticationFilter which is in charge of all
authentication.