Retry SonarQube quality gate check a few times
This should ensure that even if the webhook isn't functioning correctly,
we can still get a relatively timely finish of the pipeline.

This changes the max time to 20 minutes, with up to 4 checks with 5
minutes between each.

Note that currently we've been seeing occasional issues where the
webhook is delivered to Jenkins successfully (verified via Jenkins
logs), and yet the related job is not notified/updated.