Remove Quartz. Rely on Spring scheduling.
The using the "purge-snapshots" and "capture-snapshots" profiles, the
Spring Boot application can be started in a "one-shot" mode to run those
jobs and then quit.  Openshift's Cron features can then start pods with
those profiles active to periodically run the jobs.  In developer mode,
the jobs are run on a cron schedule set in the application properties.
If no special profiles or modes are active, the jobs do not run at all.