Ensure producer destroyed before job thread pool
Without this fix, during context destruction, first the thread pool is
destroyed, which calls interrupt on all threads (including the worker
thread which invoked `SpringApplication.exit`). The thread proceeds to
destroy the kafka producer, which calls join on one of its internal
threads. However, because `interrupt` was already invoked, an
`InterruptedException` is thrown.

With this fix, the Kafka producer is destroyed first, and so join
proceeds until the producer is done, and *then* the thread pool is
destroyed.