Add jolokia actuator
Note that I did not apply any security-related changes here. The API
gateway will already perform authn/authz for this endpoint, and since we
plan to merge rhsm-conduit and rhsm-subscriptions into a single codebase
in the future, it provides little value to do so now.

Once we merge the projects, we should apply logging of the principal as
done in RedHatInsights/rhsm-subscriptions#175.

To test try, `curl http://localhost/actuator/jolokia/list` and observe a
listing of JMX beans. Try other JMX invocations as desired.