Use generated ID instead of inventory_id for Hosts
When an ID is manually set on the Host record, spring-data
repositories will always do a merge when saving an entity.
This results in a select being executed every time a new
host is created which in not necessary since we are
deleting all hosts for an account before a tally is run.

By using a generated ID and storing the inventory_id as a
separate property, we avoid the select statements, which
improves the overall performance.