1878665: No longer throw exception collection 0 socket RHEL hypervisor
Instead of throwing an exception when collecting for a 0 socket
RHEL hypervisor, create the bucket with 0 sockets allowing it to
be filtered by the API as it currently does.

Previously, the exception was not being handled correctly leading to
an entire tally run failing for an account reporting one of these
hypervisors.