Apply origin/referer check to all jolokia requests
Even GET requests, as those can actually cause bean invocation in
Jolokia.

To be consistent with the AntiCsrfFilter, I made it so that
`DEV_MODE=true` causes the origin/referer check to be turned off.

To test, try running in the following configurations:

1. `DEV_MODE=true ./gradlew bootRun`:
   http://localhost:8080/actuator/jolokia will be fully open
2. `./gradlew bootRun`: http://localhost:8080/actuator/hawtio will be
   non-functional because /actuator/jolokia requests are rejected based
   on the default origin suffix of .redhat.com w/ port 443.
3. `RHSM_SUBSCRIPTIONS_ANTI_CSRF_DOMAIN_SUFFIX=localhost
   RHSM_SUBSCRIPTIONS_ANTI_CSRF_PORT=8080 ./gradlew`:
   http://localhost:8080/actuator/hawtio will be functional

Note that there is some nuance, as hitting `/actuator/jolokia` with `GET`
(when visiting with a browser), doesn't send an `Origin` header. So with
the restrictor turned on, visiting /actuator/jolokia is fruitless even
with the correct domain suffix/port set.