Refactor properties to be IDEA friendly
Rename rhsm-subscriptions.properties to application.properties. This
enables IDE support (autocompletion, hints) for the properties file with
no additional configuration needed. Additionally, this eliminates the
need to specify @PropertySource.

Additionally, I refactored the retention properties to be picked up via
`@Component`, as for some reason, otherwise IDEA wasn't picking it up
properly.

There seem to be some limitations with the properties support, namely
`@ConfigurationProperties` annotations on `@Beans` don't seem to be
picked up, and it seems that Map objects are not well supported. I
couldn't get a few properties to be picked up by IDEA due to these
limitations:

- `rhsm-subscriptions.package_uri_mappings.org.candlepin.subscriptions`
  because it's not implemented as a type-safe config value.
- `hawtio.*` as the hawtio properties are implemented as a `Map`
  internally.