Refactor profiles into Configuration classes
The strategy is to separate component definitions into Configuration
classes for each profile:

- `api`: `ApiConfiguration`
- `capacity-ingress`: `CapacityIngressConfiguration`
- `capture-snapshots`: `CaptureSnapshotsConfiguration`
- `liquibase-only`: `LiquibaseUpdateOnlyConfiguration`
- `worker`: `TallyWorkerConfiguration`

All of these configurations are now imported centrally into
ApplicationConfiguration and ApplicationConfiguration itself
contains only beans that are common across all profiles.

I also made a Configuration for dev-mode which was a useful way to limit
the scheduling to only when dev mode is active.

In order to limit the scope of the Configuration classes, I utilized
package scanning and composition as appropriate.

This involved some refactoring, specifically, I've eliminated the
following packages:
- `org.candlepin.subscriptions.controller` (moved controllers into
  packages by function)
- `org.candlepin.subscriptions.jackson` (moved into
  `org.candlepin.subscriptions.resteasy`)

I also refactored the packages around the in-memory queue
implementation.

The task producer/consumer configurations have been reworked so that a
component that needs to produce or consume messages should work only
with abstract configurations: `TaskProducerConfiguration` and
`TaskConsumerConfiguration`, respectively. In order to unify the
consumer interface, I had to rework the in-memory queue slightly.

For the `kafka-queue` profile, there are two configurations:

- `KafkaTaskProducerConfiguration`
- `KafkaTaskConsumerConfiguration`

I removed the need for the in-memory profile by having in-memory
configurations that are only active if kafka-queue components are
missing.

This PR also reworks the task framework slightly to allow multiple task
queues (for the merge of rhsm-conduit into this codebase).

Testing
-------
I recommend launching with various profiles and confirming expected
behavior:

- `SPRING_PROFILES_ACTIVE=api ./gradlew bootRun` should not create any
  task queue and the APIs should be accessible.
- `SPRING_PROFILES_ACTIVE=capacity-ingress ./gradlew bootRun`
  should only serve the capacity ingress API.
- `SPRING_PROFILES_ACTIVE=liquibase-only ./gradlew bootRun` should
  invoke liquibase and then exit.
- `SPRING_PROFILES_ACTIVE=capture-snapshots ./gradlew bootRun` should
  capture enqueue tasks to capture snapshots and then exit.
- `SPRING_PROFILES_ACTIVE=purge-snapshots ./gradlew bootRun` should
  cause retention logic to run and then exit.
- `SPRING_PROFILES_ACTIVE=worker ./gradlew bootRun` should consume
  tasks, but not serve the API.

Appending `kafka-queue` should cause the producers/consumers to use
kafka. This can be verified by use of the podman-compose image and
commands such as:

```
podman exec rhsm-subscriptions_kafka_1 kafka-console-consumer --bootstrap-server localhost:29092 --topic platform.rhsm-subscriptions.tasks --max-messages 1
```

NOTE: looks like the podman-compose setup uses 29092, but we defaulted
kafka to 9092... Will address in a subsequent PR. You'll need to use
KAFKA_BOOTSTRAP_PORT=29092 in the meanwhile.