Perform OpenShift metrics gathering via two PromQL queries
First pull the subscription_lables metrics that contain our metadata,
from prometheus and key the label sets by cluster_id and timestamp.

Next the OpenShift cluster cores metrics are pulled by account and an
Event object is created/persisted if a label set was found matching
the cluster ID and timestamp of the metric value.

Note:
  While this approach isn't optimal, it was done in order to get around
  a join issue caused by the original query when labels such as 'support'
  changes for a cluster.