Fix JPA mapping to address NPE during account merge
Prior to the change, if a host had tally buckets added, then they would
not receive the proper host_id populated. Apparently, it not intended to
map the `@ManyToOne` in an embedded ID. See the javadoc for `@MapsId`
for a proper example. I tried several variations; `@MapsId` isn't picked
up if applied to the `@EmbeddedId`.

I also noted that the HashMap for instances was causing extra
unnecessary update statements, so I tweaked its JPA config too.