Fix ProductProfileRegistry to find by product long form name
Initially the ProductProfileRegistry only mapped profiles against the
integer representing a product.  However, the REST interface accepts
arguments provided as the long form English name.

Currently, I have called the integer "product" and the long form names
"productIds".  Intuitively, this seems backwards but I did so to comport
with with ProductId enumeration which uses long form names.  Eventually,
I believe we should refactor the naming strategy, but that is an issue
for another day.