Fix Kafka retry code and intermittent unit test failure
The existing unit test for the Kafka retry code did not function
properly and violated some Mockito principle around its use of verify.
The result was intermittent test failures that did not even show up in
the failing test class (due to https://stackoverflow.com/a/22550055).
I did not investigate too deeply into what Mockito's actual issue
was because I realized the unit test was not actually successfully
testing the retry code.

As it was before, the retry code did not work.  Retries did not actually
occur because the KafkaException was caught immediately and did not
bubble up out of the method for the Spring Retry AOP to catch and
handle.

This patch defines the retry in a RetryTemplate that accepts a lambda
for the call we want to retry.  Only after the retries are exhausted
will the exception bubble up to the catch clause.  It also includes a
simple unit test.