Properly create/update events when processing OpenShift metrics
When metric data is being processed, update existing events and
delete any that are stale from a previous run. If the metering
process is re-run then event data should be in sync afterwards.

Prior to this change, duplicate events would be created as if
the metering process was run twice.

NOTE: If a duplicate metric is encountered for a cluster, the
last one processed will win.