Adjust event date according to average calculation range
Because we are taking the average over an hour, we need to set the event
date to the start of the hour that the query is averaging over. For
example, the query might return a metric at time of 3pm, but the value
is actually the average b/w 2pm and 3pm. We set the event timestamp to
the start 2pm and have it expire at 3pm. We adjust this by the configured
step of the query.