Remove ProductConfig from MetricUsageCollector
ProductConfig was a hard coded stand-in for a ProductProfile,
so the MetricUsageCollector was updated to operate based on a
specified ProductProfile.

The ProductProfile now has the option of defining a service type,
as well as a default SLA and Usage.

At this point we are collecting usage metrics for only OpenShift,
so the MetricUsageCollector is being created with the
OpenShiftMetrics product profile.