Fix epoch conversion to use relevant zone offset.
The code was originally written:

    LocalDateTime.ofEpochSecond(time.longValue(), 0,
        OffsetDateTime.now(zone).getOffset()).atZone(zone);

This old code takes the offset of the current timezone at the instant
the code is run.  The issue is that time zone offsets change.  Instead
the instant of time needs to be converted according to the time zone
rules that were applicable during that specific instant:

    Instant.ofEpochSecond(time).atZone(zone);

The most common change is daylight saving time, but political decisions
can also result in changes.  For example, in 2011 Samoa switched its
time zone offset. Originally, Samoa had its primary economic interaction
with the United States, but had slowly shifted to be more engaged with
Australia, so the government decided to change the timezone offset to
reflect this.  The offset went from UTC-11 to UTC+13 and Samoa skipped
the entire day of 30 December 2011 in the process.