Fix constraint violation adding metric Events
There were a couple issues with dates that needed to
be addressed. An extra hour was getting added to the
metric lookup dates when the end date was at the top
of an hour.

Also, when looking up existing dates, we need to account for
the start date shift when metrics are transformed into
Events. Because of this, we were trying to create new
Events, when we should have been updating the existing ones.