Allow for return of multiple Subscriptions from DB
Find all matching subscriptions based when we query based on a usage key
and allow MarketplaceSubscriptionIdProvider to determine which
subscription to select.  If no matches are found, call to the
MarketplaceSubscriptionCollector, save all collected subscriptions, then
select the subscription to return.

Also allow for filtering subscriptions based on date range.