Get OpenShift metering accounts from Prometheus
When OpenShift metrics collection is triggered for all accounts
(via the job or jmx), the account list is pulled from Prometheus using the
subscription_labels metric, matching on the billing_model="marketplace"
label, and will be included in the list of accounts that has
OpenShift metering enabled.

When an OpenShift metering task is executed for an account, we first
make sure that it is opted in, and then gather the meterics.