Configure jobs to run less frequently in dev mode
These schedules are only used in dev mode. There is no
need to have these run as frequenty as they are configured.
Often times it actually gets in the way of tracing logs or
modifying data. This patch reduces the frequency that they
get run and by default sets them to run early morning.