Search offering table based on role.
The product ID in the UsageCalculation.Key class is the Swatch product
id (e.g. "RHEL" or "RHEL Workstation") while the product ID in the
offering table is the engineering product ID (an integer).  Instead of
using the product ID, we will use the role(s) associated with the
product ID to search in the offering table.