Use fully qualified names for parameterized tests
The default test name for a parameterized test is just the index number
and the arguments which does not actually tell you which specific test
method is being run.  The reasoning behind this is that test runners are
expected to print the test method name and then underneath in a
tree-like fashion, print the individual instances of that test.  Our
test runner doesn't follow this behavior making it difficult to tell
what method generated a particular parameterized test.  This patch
corrects the issue by mandating a more verbose name.