Do not declare a MethodValidationPostProcessor
The Spring Core documents instruct the user to create a
MethodValidationPostProcessor in order to enable method validation.
However, Spring Boot takes care of creating that bean that itself: "The
method validation feature supported by Bean Validation 1.1 is
automatically enabled as long as a JSR-303 implementation (such as
Hibernate validator) is on the classpath" (from
https://docs.spring.io/spring-boot/docs/current/reference/htmlsingle/#boot-features-validation).

Creating our own MethodValidationPostProcessor causes
ConstraintValidator implementations to *not* receive injection from the
Spring IoC container.

It took me a very long time to figure out why my ConstraintValidator was
not being autowired, so this commit leaves a lengthy comment near the
LocalValidatorFactoryBean bean definition in hopes that future
developers inclined to declare a MethodValidationPostProcessor would do
so in that region of the code and see the warning.