Add duration validation to CaptureMetricsSnapshotTask
Validation already exists on the producer side of the
CaptureMetricsSnapshotTask; however, additional validation is needed for
messages that are both already in the queue and invalid.

Because the CaptureMetricsSnapshotTask is not instantiated through
Spring, we are invoking the validator directly.  I tried several methods
of getting the task to be Spring managed (ObjectFactory and FactoryBean
strategies) but the validator was never invoked (possibly because it is
a constructor validator).  Ideally, the task would not take the task
arguments in the constructor.  This change would allow Spring to use a
singleton approach and allow us to apply method validation.  Making this
change would require significant refactoring in the task framework,
however.