Set database to UTC during testing
A few failing tests have come up when running them locally
and were caused by what appears to be updates to various
jpa/hibernate libs. The app is run locally in UTC, and this
change ensures that the DB will be run in UTC as well. This
also mirrors how the application is currently deployed.