Add hack to commit offset on container idle
This hack uses the ability to detect when the listener is idle, along
with a utility bean: KafkaConsumerRegistry, which gives the listener
access to the raw Kafka consumer.

I set the idle timeout to 5s; I don't see a reason currently to set
differently, but I went ahead and added it as an env var (but did not
add to templates).

With this change you'll see messages right before the commits happen:

```
2021-05-04 13:49:46,306 [thread=rhsm-conduit-task-processor-0-C-1] [INFO ] [org.candlepin.subscriptions.util.SeekableKafkaConsumer] - Committing offset for [platform.rhsm-conduit.tasks-0]
2021-05-04 13:49:46,306 [thread=rhsm-subscriptions-task-processor-0-C-1] [INFO ] [org.candlepin.subscriptions.util.SeekableKafkaConsumer] - Committing offset for [platform.rhsm-subscriptions.tasks-0]
2021-05-04 13:49:46,306 [thread=openshift-metering-task-processor-0-C-1] [INFO ] [org.candlepin.subscriptions.util.SeekableKafkaConsumer] - Committing offset for [platform.rhsm-subscriptions.openshift-metering-tasks-0]
2021-05-04 13:49:46,306 [thread=marketplace-worker-0-C-1] [INFO ] [org.candlepin.subscriptions.util.SeekableKafkaConsumer] - Committing offset for [platform.rhsm-subscriptions.tally-0]
```