Add splunk sidecar config to job template
Currently we do not have access to job logs in splunk. These
logs are important when trying to determine whether a
particular process has been run for a particular account.

By configuring a Splunk sidecar for each job, it will give
us access to important logging info that will help with debugging
issues.