ENT-2664 Add logic to convert upstream products into an Offering
The mechanism that performs the conversion is called MidProduct, because
it sits in the middle between the product types from upstream and the
Offering type of Swatch.

MidProduct is in charge of knowing which product trees and eng products
to pull in order to return an Offering.

The testing for MidProduct+OfferingSyncController is done with
OfferingSyncControllerTest. It uses a specific set of actual SKUs and
engOIDs, stripped down to only the data used by Swatch, to test normal
cases and corner cases. Examples of corner cases are SKUs with derived
products or SKUs with unlimited cores or sockets.