Added a PromQL query builder
The QueryBuilder uses Spring Expression Language (SPEL) to apply
variables to a configured query template.

A query template is expected to be a well formed PromQL query.
Query templates are to be defined as a map (templateKey: promql)
in a configurationfile and can be be accessed via
PrometheusMetricsProperties:

   rhsm-subscriptions.metering.prometheus.metric.queryTemplates

A template can include template variables via a SPEL expressions.
These variables are substituted when the query is built and template
values are defined in the QueryDescriptor passed to the builder
at build time. Every property defined in the descriptor object will
accessible as template variables when the query is built. Variable
names are specified in dot notation and can be nested to access any
value defiend in the descriptor.

A QueryDescriptor contains two key fields that can be accessed in
a query template.

  tag: Comes from a configured TagMetric
  runtime: A generic collection of runtime args passed at runtime.

While the example below is not PromQL, an example of accessing
descriptor values would be as follows:

  Metric #{tag.metricId} looked up for Account #{runtime[account]}

NOTE:
Template variables can contain other SPEL expressions, but it is important
note that we support 3 levels of nesting out of the box. This is not a
supported feature of SPEL. The number of levels we support was made
configurable, but I can't see us needing to change it.